// types/zod/inputTypeSchemas/driversUncheckedCreateInputSchema.ts
export const driversUncheckedCreateInputSchema: z.ZodType<Prisma.driversUncheckedCreateInput> = z
	.object({
		driver_id: z.string().uuid().optional(),
		business_id: z.string().optional().nullable(),
		online: z.boolean().optional().nullable(),
		on_order: z.boolean().optional().nullable(),
		working_hours: z.union([z.lazy(() => NullableJsonNullValueInputSchema), InputJsonValueSchema]).optional(),
		ride_requirements: z.union([z.lazy(() => NullableJsonNullValueInputSchema), InputJsonValueSchema]).optional(),
		created_at: z.coerce.date().optional(),
		updated_at: z.coerce.date().optional(),
		user_id: z.string().optional().nullable(),
		current_vehicle_id: z.string().optional().nullable(),
		last_used_vehicle_id: z.string().optional().nullable(),
		location: z.union([z.lazy(() => NullableJsonNullValueInputSchema), InputJsonValueSchema]).optional(),
		delivery_timeline: z.union([z.lazy(() => NullableJsonNullValueInputSchema), InputJsonValueSchema]).optional(),
		last_ping_at: z.coerce.date().optional(),
		is_inactive: z.boolean().optional().nullable(),
		transfer_requirements: z
			.union([z.lazy(() => NullableJsonNullValueInputSchema), InputJsonValueSchema])
			.optional(),
		regions: z.union([z.lazy(() => driversCreateregionsInputSchema), z.string().array()]).optional(),
		handles_taxi_orders: z.boolean().optional().nullable(),
		handles_transfer_orders: z.boolean().optional().nullable(),
		handles_delivery_orders: z.boolean().optional().nullable(),
		taxi_orders_toggled: z.boolean().optional().nullable(),
		transfer_orders_toggled: z.boolean().optional().nullable(),
		delivery_orders_toggled: z.boolean().optional().nullable(),
		partner_cash_balance: z.number().optional().nullable(),
		come_to_work_last_sent_at: z.coerce.date().optional().nullable(),
		vehicles: z.lazy(() => vehicle_driversUncheckedCreateNestedManyWithoutDriverInputSchema).optional(),
		orders: z.lazy(() => taxi_ordersUncheckedCreateNestedManyWithoutDriverInputSchema).optional(),
		received_orders: z.lazy(() => taxi_order_sentUncheckedCreateNestedManyWithoutDriverInputSchema).optional(),
		delivery_orders: z.lazy(() => delivery_ordersUncheckedCreateNestedManyWithoutDriverInputSchema).optional(),
		received_delivery_orders: z
			.lazy(() => delivery_order_sentUncheckedCreateNestedManyWithoutDriverInputSchema)
			.optional(),
		documents: z.lazy(() => documentsUncheckedCreateNestedManyWithoutDriversInputSchema).optional(),
		driver_history_locations: z
			.lazy(() => driver_history_locationsUncheckedCreateNestedManyWithoutDriverInputSchema)
			.optional(),
		driver_municipalities: z
			.lazy(() => driver_municipalitiesUncheckedCreateNestedManyWithoutDriversInputSchema)
			.optional(),
		activity_logs: z.lazy(() => driver_activity_logsUncheckedCreateNestedManyWithoutDriverInputSchema).optional(),
	})
	.strict();